Refactor MCP module to mcp_client and update imports
Renamed the openhands.mcp module and its files to openhands.mcp_client for clarity and separation of concerns. Updated all relevant imports throughout the codebase and tests to reference openhands.mcp_client.
