+++ openhands/mcp_client/__init__.py
@@ -1,7 +1,7 @@
-from openhands.mcp.client import MCPClient
-from openhands.mcp.error_collector import mcp_error_collector
-from openhands.mcp.tool import MCPClientTool
-from openhands.mcp.utils import (
+from openhands.mcp_client.session import MCPClient
+from openhands.mcp_client.error_collector import mcp_error_collector
+from openhands.mcp_client.tool import MCPClientTool
+from openhands.mcp_client.utils import (
     add_mcp_tools_to_agent,
     call_tool_mcp,
     convert_mcp_clients_to_tools,
