+++ openhands/mcp_client/error_collector.py
