+++ openhands/runtime/impl/action_execution/action_execution_client.py
@@ -456,8 +456,8 @@ class ActionExecutionClient(Runtime):
             return ErrorObservation('MCP functionality is not available on Windows')
 
         # Import here to avoid circular imports
-        from openhands.mcp.utils import call_tool_mcp as call_tool_mcp_handler
-        from openhands.mcp.utils import create_mcp_clients
+        from openhands.mcp_client.utils import call_tool_mcp as call_tool_mcp_handler
+        from openhands.mcp_client.utils import create_mcp_clients
 
         # Get the updated MCP config
         updated_mcp_config = self.get_mcp_config()
