+++ tests/unit/test_mcp_error_collector.py
@@ -2,7 +2,7 @@
 
 import time
 
-from openhands.mcp.error_collector import (
+from openhands.mcp_client.error_collector import (
     MCPError,
     MCPErrorCollector,
     mcp_error_collector,
