+++ tests/unit/test_mcp_timeout.py
@@ -4,7 +4,7 @@ from unittest import mock
 import pytest
 
 from openhands.core.config.mcp_config import MCPConfig, MCPSSEServerConfig
-from openhands.mcp import MCPClient, create_mcp_clients, fetch_mcp_tools_from_config
+from openhands.mcp_client import MCPClient, create_mcp_clients, fetch_mcp_tools_from_config
 
 
 @pytest.mark.asyncio
@@ -22,7 +22,7 @@ async def test_sse_connection_timeout():
     mock_client.disconnect = mock.AsyncMock()
 
     # Mock the MCPClient constructor to return our mock
-    with mock.patch('openhands.mcp.utils.MCPClient', return_value=mock_client):
+    with mock.patch('openhands.mcp_clientutils.MCPClient', return_value=mock_client):
         # Create a list of server URLs to test
         servers = [
             MCPSSEServerConfig(url='http://server1:8080'),
@@ -50,7 +50,7 @@ async def test_fetch_mcp_tools_with_timeout():
     mock_config.shttp_servers = []
 
     # Mock create_mcp_clients to return an empty list (simulating all connections failing)
-    with mock.patch('openhands.mcp.utils.create_mcp_clients', return_value=[]):
+    with mock.patch('openhands.mcp_clientutils.create_mcp_clients', return_value=[]):
         # Call fetch_mcp_tools_from_config
         tools = await fetch_mcp_tools_from_config(mock_config, None)
 
@@ -88,7 +88,7 @@ async def test_mixed_connection_results():
 
     # Mock create_mcp_clients to return our successful client
     with mock.patch(
-        'openhands.mcp.utils.create_mcp_clients', return_value=[successful_client]
+        'openhands.mcp_clientutils.create_mcp_clients', return_value=[successful_client]
     ):
         # Call fetch_mcp_tools_from_config
         tools = await fetch_mcp_tools_from_config(mock_config, None)
